Move newGithubClient to ghclient.go
